requirements.txt
discord.py
pyperclip
python-dotenv
